ai/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
fastapi==0.100.0
uvicorn==0.23.1
python-dotenv==1.0.0
tensorflow==2.13.0
pymongo==4.5.0
pytest==7.4.2
black==23.7.0
flake8==6.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 